+++ src/sites/mobile-taro/vue/src/base/pages/button/index.vue
@@ -31,28 +31,20 @@
     <div class="demo-button-row2">
       <nut-button loading type="info"></nut-button>
       <nut-button loading type="warning">加载中...</nut-button>
-      <nut-button :loading="isLoading" type="success" @click="changeLoading"
-        >Click me!</nut-button
-      >
+      <nut-button :loading="isLoading" type="success" @click="changeLoading">Click me!</nut-button>
     </div>
     <h2>图标按钮</h2>
     <div class="demo-button-row2">
-      <nut-button
-        shape="square"
-        plain
-        type="primary"
-        icon="star-fill"
-      ></nut-button>
+      <nut-button shape="square" plain type="primary" icon="star-fill"></nut-button>
       <nut-button shape="square" type="primary" icon="star">收藏</nut-button>
     </div>
 
     <h2>按钮尺寸</h2>
     <div class="demo-button-row2">
-      <nut-button size="large" type="primary" style="margin-bottom: 10px"
-        >大号按钮</nut-button
-      >
+      <nut-button size="large" type="primary" style="margin-bottom: 10px">大号按钮</nut-button>
       <nut-button type="primary">普通按钮</nut-button>
       <nut-button size="small" type="primary">小型按钮</nut-button>
+      <nut-button size="mini" type="primary">迷你按钮</nut-button>
     </div>
     <h2>块级元素</h2>
     <div class="demo-button-row2">
@@ -62,9 +54,7 @@
     <div class="demo-button-row2">
       <nut-button color="#7232dd">单色按钮</nut-button>
       <nut-button color="#7232dd" plain>单色按钮</nut-button>
-      <nut-button color="linear-gradient(to right, #ff6034, #ee0a24)">
-        渐变按钮
-      </nut-button>
+      <nut-button color="linear-gradient(to right, #ff6034, #ee0a24)"> 渐变按钮 </nut-button>
     </div>
   </div>
 </template>